Update 06_Lucky_Uniccorn.py
Thanks for this.  I'd like to suggest a small change to your code.  I've edited the if block from line 79 on-wards so that the balance is included as part of the feedback.  If the user has $0.50, it tells them they don't have enough to play.  Your program was similar but the balance was given before the user was told they could not continue - it made sense to combine those statements.

I also edited the condition for the while loop so that if the amount falls below $1, the game will end.  This means we don't need to set keep_going to 'end' when the users has run out of money leading to slightly more efficient code.